Check if child exists for current user (parent) before updating or gotostory page

+++ src/RW/Modules/InnerHashTagContent.php
@@ -3,6 +3,7 @@
 namespace RW\Modules;
 
 use RW\Modules\HashTagContent;
+use RW\Modules\Child;
 
 class InnerHashTagContent extends HashTagContent
 {
@@ -10,22 +11,28 @@ class InnerHashTagContent extends HashTagContent
     public static function goToStory(array $data)
     {
         if ( (string) __FUNCTION__ === (string) $data['action']) {
-            echo 'Story';
+            if ( Child::exists($data['id']) ) {
+                echo 'Story';
+            }
         }
     }
 
     public static function editInfo(array $data)
     {
         if ( (string) __FUNCTION__ === (string) $data['action']) {
-            self::create(true, $data['id']);
+            if ( Child::exists($data['id']) ) {
+                self::create(true, $data['id']);
+            }
         }
     }
 
     public static function deleteChild(array $data)
     {
         if ( (string) __FUNCTION__ === (string) $data['action']) {
-            if (wp_trash_post($data['id'])) {
-                echo json_encode(array('status' => 'Child deleted successfully.'));
+            if ( Child::exists($data['id']) ) {
+                if (wp_trash_post($data['id'])) {
+                    echo json_encode(array('status' => 'Child deleted successfully.'));
+                }
             }
         }
     }
